Refactored file Structure to DDD
