+++ src/auth/auth.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { AuthService } from './services/auth.service';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { JwtStrategy } from './jwt.strategy';
 import { UserModule } from '../user/user.module';
-import { UserRepository } from 'src/user/user.repository';
+import { UserRepository } from 'src/user/repositories/user.repository';
 import { MongooseModule } from '@nestjs/mongoose';
-import { UserSchema, User } from 'src/user/user.schema';
-import { AuthController } from './auth.controller';
+import { UserSchema, User } from 'src/user/schemas/user.schema';
+import { AuthController } from './controllers/auth.controller';
 
 @Module({
   imports: [
