+++ src/user/controllers/user.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Post, Body } from '@nestjs/common';
-import { UserService } from './user.service';
+import { UserService } from '../services/user.service';
 
 @Controller('user')
 export class UserController {
