+++ src/user/user.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { User, UserSchema } from './user.schema';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
-import { UserRepository } from 'src/user/user.repository';
+import { User, UserSchema } from './schemas/user.schema';
+import { UserService } from './services/user.service';
+import { UserController } from './controllers/user.controller';
+import { UserRepository } from 'src/user/repositories/user.repository';
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
